requirements.txt
fastapi~=0.54.2
sqlalchemy~=1.3.17
starlette~=0.13.2
alembic~=1.4.2
psycopg2~=2.8.5
python-multipart~=0.0.5
python-dotenv~=0.13.0
aiofiles~=0.5.0
